// src/features/access-control/types/access-control.types.ts
export interface FullUserDto {
  id: number;
  firstName?: string;
  lastName?: string;
  fullName?: string;
  email?: string;
}
// ...
export interface VisibilityPolicyDto {
  id: number;
  createdDate: string;
  updatedDate?: string;
  deletedDate?: string;
  isDeleted: boolean;
  createdByFullUser?: FullUserDto;
  updatedByFullUser?: FullUserDto;
  deletedByFullUser?: FullUserDto;
  code: string;
  name: string;
  entityType: string;
  description?: string;
  scopeType: number;
  includeSelf: boolean;
  isActive: boolean;
  priority?: number;
}